+++ service/controller/v13/resource/instance/create_deployment_initialized.go
@@ -1,0 +1,50 @@
+package instance
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/giantswarm/azure-operator/service/controller/v13/key"
+	"github.com/giantswarm/azure-operator/service/controller/v13/resource/instance/internal/state"
+	"github.com/giantswarm/microerror"
+)
+
+func (r *Resource) deploymentInitializedTransition(ctx context.Context, obj interface{}, currentState state.State) (state.State, error) {
+	customObject, err := key.ToCustomObject(obj)
+	if err != nil {
+		return currentState, microerror.Mask(err)
+	}
+	deploymentsClient, err := r.getDeploymentsClient(ctx)
+	if err != nil {
+		return currentState, microerror.Mask(err)
+	}
+
+	d, err := deploymentsClient.Get(ctx, key.ClusterID(customObject), key.VmssDeploymentName)
+	if IsDeploymentNotFound(err) {
+		r.logger.LogCtx(ctx, "level", "debug", "message", "deployment not found")
+		r.logger.LogCtx(ctx, "level", "debug", "message", "waiting for creation")
+		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
+		return currentState, nil
+	} else if err != nil {
+		return currentState, microerror.Mask(err)
+	}
+
+	s := *d.Properties.ProvisioningState
+	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deployment is in state '%s'", s))
+
+	if !key.IsSucceededProvisioningState(s) {
+		r.debugger.LogFailedDeployment(ctx, d, err)
+
+		if key.IsFinalProvisioningState(s) {
+			// Deployment is not running and not succeeded (Failed?)
+			// This indicates some kind of error in the deployment template and/or parameters.
+			// Restart state machine on the next loop to apply the deployment once again.
+			// (If the azure operator has been fixed/updated in the meantime that could lead to a fix).
+			return DeploymentUninitialized, nil
+		} else {
+			return currentState, nil
+		}
+	}
+
+	return ProvisioningSuccessful, nil
+}
