+++ service/controller/v13/resource/instance/create_provisioning_successful.go
@@ -1,0 +1,12 @@
+package instance
+
+import (
+	"context"
+
+	"github.com/giantswarm/azure-operator/service/controller/v13/resource/instance/internal/state"
+)
+
+func (r *Resource) provisioningSuccessfulTransition(ctx context.Context, obj interface{}, currentState state.State) (state.State, error) {
+	r.logger.LogCtx(ctx, "level", "debug", "message", "vmss deployment successful")
+	return InstancesUpgrading, nil
+}
