+++ service/controller/v13/resource/instance/internal/state/error.go
@@ -1,0 +1,15 @@
+package state
+
+import "github.com/giantswarm/microerror"
+
+// executionFailedError is an error type for situations where Resource
+// execution cannot continue and must always fall back to operatorkit.
+//
+// This error should never be matched against and therefore there is no matcher
+// implement. For further information see:
+//
+//     https://github.com/giantswarm/fmt/blob/master/go/errors.md#matching-errors
+//
+var executionFailedError = &microerror.Error{
+	Kind: "executionFailedError",
+}
