+++ service/controller/v13/resource/instance/status.go
@@ -15,10 +15,11 @@ const (
 	DeploymentParametersChecksum = "ParametersChecksum"
 
 	// States
-	DeploymentInitialized  = "DeploymentInitialized"
-	DeploymentCompleted    = "DeploymentCompleted"
-	InstancesUpgrading     = "InstancesUpgrading"
-	ProvisioningSuccessful = "ProvisioningSuccessful"
+	DeploymentUninitialized = "DeploymentUnitialized"
+	DeploymentInitialized   = "DeploymentInitialized"
+	DeploymentCompleted     = "DeploymentCompleted"
+	InstancesUpgrading      = "InstancesUpgrading"
+	ProvisioningSuccessful  = "ProvisioningSuccessful"
 )
 
 func (r *Resource) deleteResourceStatus(customObject providerv1alpha1.AzureConfig, t string, s string) error {
